Move tooltip to separate plugin
